+++ app_cpu.py
@@ -23,12 +23,14 @@ def update_board(an_array):
   return a_new_array
 
 def conway(HEIGHT, WIDTH, ITERATIONS, an_array):
-  time.sleep(0.5)
-  dpg.delete_item("viewport_back", children_only=True)
+  if (not RUNTIME_FLAG):
+    time.sleep(0.5)
+    dpg.delete_item("viewport_back", children_only=True)
 
-  while dpg.is_dearpygui_running():
-    for _ in range(ITERATIONS):
-      an_array = update_board(an_array)
+  for _ in range(ITERATIONS):
+    an_array = update_board(an_array)
+
+    if (not RUNTIME_FLAG):
       for i in range(WIDTH):
         for j in range(HEIGHT):
           value = an_array[i, j]
@@ -36,30 +38,32 @@ def conway(HEIGHT, WIDTH, ITERATIONS, an_array):
             dpg.draw_rectangle((i * PIXEL_DIM[0], j * PIXEL_DIM[1]), ((i * PIXEL_DIM[0]) + PIXEL_DIM[0], (j * PIXEL_DIM[1]) + PIXEL_DIM[1]), fill = (255, 255, 255, 255), parent = "viewport_back")
           else:
             dpg.draw_rectangle((i * PIXEL_DIM[0], j * PIXEL_DIM[1]), ((i * PIXEL_DIM[0]) + PIXEL_DIM[0], (j * PIXEL_DIM[1]) + PIXEL_DIM[1]), fill = (0, 0, 0, 0), parent = "viewport_back")
-      
+
       dpg.render_dearpygui_frame()
       time.sleep(0.5)
       dpg.delete_item("viewport_back", children_only=True)
-    break
 
-  dpg.destroy_context()
+  if (RUNTIME_FLAG):
+    END_TIME = time.time()
+    print('Execution time: ', END_TIME - START_TIME)
 
 def create_board(HEIGHT, WIDTH, ITERATIONS):
-  dpg.setup_dearpygui()
-  dpg.add_viewport_drawlist(front=False, tag="viewport_back")
-  dpg.show_viewport()
-
   an_array = np.asmatrix([[random.randint(0, 1) for _ in range(HEIGHT)] for _ in range(WIDTH)])
 
-  for i in range(WIDTH):
-    for j in range(HEIGHT):
-      value = an_array[i, j]
-      if value == 0:
-        dpg.draw_rectangle((i * PIXEL_DIM[0], j * PIXEL_DIM[1]), ((i * PIXEL_DIM[0]) + PIXEL_DIM[0], (j * PIXEL_DIM[1]) + PIXEL_DIM[1]), fill = (255, 255, 255, 255), parent = "viewport_back")
-      else:
-        dpg.draw_rectangle((i * PIXEL_DIM[0], j * PIXEL_DIM[1]), ((i * PIXEL_DIM[0]) + PIXEL_DIM[0], (j * PIXEL_DIM[1]) + PIXEL_DIM[1]), fill = (0, 0, 0, 0), parent = "viewport_back")
-  
-  dpg.render_dearpygui_frame()
+  if (not RUNTIME_FLAG):
+    dpg.setup_dearpygui()
+    dpg.add_viewport_drawlist(front=False, tag="viewport_back")
+    dpg.show_viewport()
+
+    for i in range(WIDTH):
+      for j in range(HEIGHT):
+        value = an_array[i, j]
+        if value == 0:
+          dpg.draw_rectangle((i * PIXEL_DIM[0], j * PIXEL_DIM[1]), ((i * PIXEL_DIM[0]) + PIXEL_DIM[0], (j * PIXEL_DIM[1]) + PIXEL_DIM[1]), fill = (255, 255, 255, 255), parent = "viewport_back")
+        else:
+          dpg.draw_rectangle((i * PIXEL_DIM[0], j * PIXEL_DIM[1]), ((i * PIXEL_DIM[0]) + PIXEL_DIM[0], (j * PIXEL_DIM[1]) + PIXEL_DIM[1]), fill = (0, 0, 0, 0), parent = "viewport_back")
+    
+    dpg.render_dearpygui_frame()
   conway(HEIGHT, WIDTH, ITERATIONS, an_array)
 
 parser = argparse.ArgumentParser(description = "Initialize board dimensions")
@@ -75,6 +79,11 @@ ITERATIONS = args["iterations"]
 RUNTIME_FLAG = args["runtime"]
 PIXEL_DIM = (16, 16)
 
-dpg.create_context()
-dpg.create_viewport(title='Conway\'s Game of Life: CPU', width=WIDTH*PIXEL_DIM[0], height=HEIGHT*PIXEL_DIM[1]+50, resizable = False)
+if (not RUNTIME_FLAG):
+  dpg.create_context()
+  dpg.create_viewport(title='Conway\'s Game of Life: CPU', width=WIDTH*PIXEL_DIM[0], height=HEIGHT*PIXEL_DIM[1]+50, resizable = False)
+
+if (RUNTIME_FLAG):
+  START_TIME = time.time()
+
 create_board(HEIGHT, WIDTH, ITERATIONS)
